Eilmer3: Added viscous_signal_factor for attenuating the viscous contribution to the signal-speed for the calaculation of a stable time step.
The default value of 1.0 will apply the full viscous effect for the signal
calculation within the time-step calculation.  It has been suggested that
the full viscous effect may not be needed to ensure stable calculations
for highly-resolved viscous calculations.  A value of 0.0 will completely
suppress the viscous contribution to the signal speed calculation.