Eilmer4: User-defined source terms and MMS case added.
The user-defined source terms are now available
in the simulation code. To test this, I reinstated the
manufactured solution case for the simple Euler flow.
The L2 norm in density is:
  0.013370302107

I'm going to claim this compares well enough to the
eilmer3 value of:
  0.0132608184103

Of course, the real test is that the order of spatial convergence
still holds. I should do that test soon.

Those differences could be due to representation precision
in all of the input/output, or (the bigger unknown) differences
in how the compilers have optimised certain expressions.