lib/gas/kinetics: Support for Troe-form of pressure-dependent rate when F_cent given.
The Troe-form for pressure dependent rate is usually given in terms
of four parameters: a , T*, T**, and T***. F_cent is computed as
a function of these parameters. In some cases, F_cent is just given
as a constant. I gather that the Chemkin way to deal with this
is just to set T*, T** and T*** to very high or very low values (like
1.0e30 or 1.0e-30) so that they have no real effect.

In our implementation, if F_cent is supplied then it is assumed
constant and used directly. There is no need to trick the
implementation with values for the other parameters.  The code
will not look for the other parameters if it finds F_cent.