eilmer3: Added a check_volumes method to Block3D class.
This allows one to perform a check of cell volumes
at the preparation stage. This is much the same
as the functionality added in Rev. 1793 that
handled the 2D case.

To add the volume check in a 3D eilmer setup,
add something like the following to your
Python input script:

for blk in Block.blockList:
    print "Checking block: ", blk.blkId
    bad_cells = blk.check_volumes(gdata)
    if bad_cells:
        for c in bad_cells:
            print "Bad cell (i,j,k,vol): ", c[0], c[1], c[2], c[3]