Eilmer4: ported example of sharp-nosed 2D body in Mach 3 flow.
Made a couple of small code changes to get this case running.
(1) Blocks must have labels for the postprocessor.
(2) LuaFnPath is a valid Path type.

The clustering functions need to be propagated through to the grid generator
to complete this example but, at least, it does now run on uniform grids.