eilmer3: Added "before_grid_update" user-defined function hook.
An additionaly user-defined hook has been added called:
"before_grid_update".
This is called, as the name implies, before the grid movement
step in a moving grid simulation. If the function is not present,
then the call is simply skipped. It is enabled in the same way as
the "at_timestep_start" and "at_timestep_end" user-defined functions:
by setting a udf_file in the input script.