Eilmer4: make use of some parallelism at initialization.
Just exploring how difficult it is going to be to get the code running in
parallel across the collection of blocks.  Some parts of the loops have been
easy to run in parallel but there are a couple of calls that cause segmentation
faults if run in parallel.  The comments mark these.

This commit should build and run reliably, despite the experiment parallel loops.