Eilmer4: have removed the parallel loops inserted in rev 2148.
Parallel loops at initialization broke the code in a subtle way.
The number of steps for the cone20-simple remained correct but the flow field
looked more like a 2D-planar flow.  Strange.

If I had put the fully automated tests in place, this broken code would not
have slipped past because the shock shape would have been checked as well as
the number of steps taken by the simulation.  There's a lesson in that...