Eilmer4: Try to be clearer with grid-time-level and flow-time-level indices.
To accommodate a multi-stage update for flow quantities and allow for
a moving grid, there is both a flow-time_level (ftl) index and a grid-time_level
(gtl) index for accessing the data.  In a fixed-grid simulation gtl is always 0.

Some functions need only a gtl index and some need both a gtl and ftl index.
I have tried to be consistent and always put the gtl index first.  I think
that there are no functions taking just a flow-time-level index.