Eilmer4: calculation of the secondary-cell area is now done by gradient_xy().
This will add some extra computation per time step but the advantage is that
we can dispose of the function that computes secondary-cell areas and not have
to worry about whether we have consistent definitions of teh secondary cells,
at least for 2D flows.  We have yet to port and refactor the code for 3D flow
gradients.