dlang/gas + Eilmer4 + kinetics: Move static work arrays to private local arrays.
To make the code a little friendlier to doing block-level
parallelism in a shared memory manner, we are attempting
to push all work arrays as private class data. When the work
arrays are static, it's difficult (for us) to reason
when the local thread memory is allocated.

Using private arrays means that the 'const' modifier is
inappropriate: we can no longer guarantee that an object's
state will not be changed during a method invocation.
In fact, the working arrays are likely to be modified
on many method invocations.

Removing the 'const' modifier had a ripple effect throughout
the code, hence the large changeset in this commit.