Eilmer4: Put references to defining vertices into FVInterface objects.
This allows the viscous flux calculation to proceed with mostly "local" data.
If we can get more SBlock functions to work on local data, without having to
know the details of the underlying grid, then we can eventually move those
functions to the Block base class.  Having to maintain less grid-specific code
will be good.