Eilmer4: computations on each SBlock should now be independent.
To achieve this independence, each block has it's own GasModel object and
a copy of any work arrays (used in the flow field interpolation and the like).
This should allow reliable parallel computation with the parallel work-unit
being the block.  We won't try to go any finer than that.

Not that, for the initialization (when blocks and boundary conditions are being
constructed from their JSON descriptions), there is still a gmodel available in
the GlobalConfig class.  This gmodel_master is also used in the Lua interface,
which should only ever be actively used in a serial mode, I think.