Eilmer4: Fixed bug in setting vel.y in user-defined interface b.c.
With this bug fix, the viscous version of the Manufactured Solution
case appears to behave as expected.

I also needed to pass through the gas model as an argument at
the construction stage of the boundary conditions. It seemed
that trying to pull the reference from the parent block
did not work since the block itself was not completely constructed.