Eilmer4: finished making dedicated Config space for the blocks/threads.
The bad news is that the cone20-simple case did not speed up with the
application of more cores.  On helmholtz, it nows runs 850 steps in 25 seconds.
But this is a small job with tow blocks only and most processing in one block.
It's a poor fit for a 4-core machine so, maybe, the overhead continues to be a
problem.

The good news is that the forward-facing-step problem runs in the improved time
of 484 seconds, compared with 556 seconds for the serial version of a few code
revisions back.  This also, is not well load balanced and has only 3 blocks
with 76% of total cells in the largest block.  With no overheads, the best
wall-clock time we could reasonable expect is then 423 seconds.  Looks like our
parallel code is doing OK, especially since we aren't doing the boundary effects
in parallel, yet.