Eilmer4: the parallel version is good! (I think).
The poor timing for the cone20-simple case that was noted in the last commit
turned out to be a matter of not using a multicore machine appropriately.
Until this commit, all tests were run using the dedfault number of threads in
the thread pool for parallel loops.  On my 4 core AMD machine, totalCPUs is
reported as 4 but I was running a simulation with only 2 blocks and it seems
that too many cores were spoiling the run time.  Here are the times now that
we can specify the number of CPUs to use.

850 steps of the cone20-simple case
max-cpus  WallClock seconds
1         20  just as good as the serial simulation ever was!
2         17  the second cpu is doing some work but one block is small
3         20  the third cpu is wasted, and actually makes the job slower
4         25  the fourth cpu makes things even slower

1500 steps of the forward-facing-step case
max-cpus  WallClock seconds
1         544  as good as previous serial simulations (difference in the noise)
3         460  nice, the best we could hope for is 423 because of the blocks

This is a pleasing result because we have yet to squeeze some more performance
out of doing the boundary conditions in parallel.