Eilmer4: call applyPreSpatialDerivAction only from a serial loop.
We haven't yet made the boundary-condition effects safe for parallel application.
This is on the [TODO] list but we can get a fair bit of benefit from having the
rest of the gas-dynamic updates done in parallel.

As I write this message, the cyl50 case is running on 4 cpus and showing a lot
of speed-up.