Eilmer4: forward-facing step case load-balanced by breaking it into more blocks.
This makes good use of SBlockArray and improves the parallel run time from 460
seconds to 299 seconds.  There's still more to gain by getting the boundary
conditions to be applied in parallel.  Ideally, we should be able to get down
to about 200 seconds run time on helmholtz.  We expect it to be somewhat over
200 seconds because we won't be initializing and writing files in parallel.
Anyway, we'll see...