Eilmer4: do the initializaion og gas blocks in phases.
This is working toward being able to embed block references into the
individual boundary conditions.  We don't have valid references until
the blocks are fully constructed and, for exchange boundary conditions,
we need references to other blocks in the collection.

The end goal is to allow application of boundary conditions in parallel
(over the block collection, that is).