Eilmer4: embed SBlock references into the boundary condition objects.
This allows us to apply the BC effects in parallel, I think...
Actually trying that is going to have to wait until the next commit.
The code still runs with serial application of the BCs across blocks.

The blocking issue is that the associative array of Lua interpreters
in the user-defined boundary condition effect objects need to be replaced
with the single Lua interpreter belonging to the block object.