Eilmer4: user-defined effects now work from the SBlock-owned LuaState.
Notes to RJG:
This simplifies the code somewhat because the one Lua interpreter is always
initialized and is used for the source terms and all of the boundary effects.
The names of the user defined functions are now of the form ghostCells_xxxx
and interface_xxxx where xxxx is one of the boundary face names.
This eliminates the need to the whichBoundary global variable to be set
before calling the user-defined Lua functions.

There's a fly in the ointment, in that the very convenient sampleFlow function
will only work within user-defined functions that are called up from serial
foreach loops in the D-domain.  This is because the function needs to dip into
the gasBlocks array to get at the flow data and this array will only be correctly
initialized for the main thread of a parallel loop.  The other htreads will
probably have null entries for the individual SBlock references.  Maybe there's
a fix that will present itself...