Eilmer4: apply boundary effects in parallel, if appropriate.
Some of the user-defined boundary effects must not be done in parallel so, for
those cases, config.apply_bcs_in_parallel may be set as false.
Run time for the forward-facing-step case decreased to 275 seconds on 3 cores
of helmholtz.  (Down from 299 seconds on 11-May.)
Run time for cone20-simple didn't change at all -- too much overhead on a small
calculation, I presume.
The manufactured solution cases need to be run with --max-cpus=1.