Eilmer4 & support modules: Removed all LuaD dependency.
This is a rather large change, and not fully tested. In this commit,
I have removed all dependency on LuaD. The cone20-udf test case runs.
I expect there will be a few more bugs introduced during the refactoring
of this code, that I will find as I proceed with testing.