Eilmer3/4: Fixed bugs in manufacture sources terms for Euler case.
I introduced this bug back in rev 1631 when I changed to using sympy.
I had tested the Navier-Stokes case and was satisfied, but neglected
to run the Euler case. It turns out, I had not set viscosity
and thermal conductivity to zero when generating the source terms
for the Euler equations.