dlang/geom: Simplify Path classes.
We shall do subranging and arc-length parameterization as derived classes.
This allows us to get rid of the t0,t1 subranging code and the arc-length
lookup code from within the basic Path classes, thus simplifying them a lot.
When we want the modified Paths, we'll build composite Path objects,
with an underlying basic Path object and the extra bits added, as needed.