Eilmer4: fiddle around with an ideal-air model for the n90-cylinder simulation.
It runs fine with a single-species ideal-air model but still does not behave
with the two-species nitrogen model.  With this commit, the input script results
in a calculation that stalls a few tenths of a microsecond into the simulation
with one cell displaying crazy-high velocities.  There are other disturbing
features in the flow solution...
We are not yet finished with debugging this case.