Eilmer4 development: Currently working on fixing the two-species n90 case.
This commit eliminates a segfault when mass-fraction arrays of differing length
are not bound-checked and the exception not caught.

At this point the two-species, non-reacting simulation behaves badly for the
optimized, release build.  The simulation grinds to a halt with very small
time step and a cell with clearly bad data.  Unfortunately for me, the debug
build runs the same simulation just fine and produces what looks to be
the correct result.  This is going to require more thought...