Eilmer4: Fix FullFaceExchangeBC to update interface properties at edges.
In prep.lua, we specify the actions of our composite boundary conditions.
In this bug fix, the FullFaceExchangeBC is now set to update the interface
properties as a pre-spatial derivative action. This relies on the fact
that some of the interface properties are set earlier at the reconstruction
stage of the convective flux update. This pre-spatial derivative action then
ensures that the rest of the properties are up-to-date.