Eilmer4: Fixed bug in post-process related to finding variable names for norms.
Since the list of variable names can be different in gas and solid domains,
we need to treat the search for names carefully. The idea is that when a name
can't be found in a certain class of flow block, just to skip the name but
report to the user that it has been skipped. Rather than skipping entries,
my bug was that I was completely 'break'ing out of the loop and so ignoring
subsequent variable names in the list. This has been corrected by using a
'continue' statement instead.