Eilmer4: Added a boundary interface effect that sets temperature at gas/solid interface.
I had previously only set the energy flux at a gas/solid interface.
However, we do require the temperature at a gas interface prior to computing
the spatial derivatives. This is because the temperature is used to compute
the viscosity at the interface, and the viscosity is used in computing the
momentum flux at that interface.

The high-level boundary condition in prep.lua now takes care of handling
this detail.