Eilmer4: small edits to get the code to compile with gdc and ldc2.
These alternate compilers are currently up to the dmd front-end 2.066.1
so a few small items in the code needed to be adjusted.  There should be
no changes to functionality.  The reason for getting the code to run with
the alternate compilers is to allow the use external tools such as valgrind
in a meaningful way.  Word is that valgrind gives lots of false positives
for dmd-compiled code.

An interesting point is that, running the n90 reacting case on helmholtz
and making 4 CPUs available for threading (needlessly, I must confess)
results in a run time of 635 seconds for the dmd 2.068.0 compiled code and
a run time of 638 seconds for the ldc2-compiled code.  So much for the usual
internet wisdom that dmd produces slow code.  Also, when limiting the number
of available CPUs to 1 for our single-block calculation, the run time for the
dmd-compiled code was 631 seconds.  So, the overheads of the parallel loops
seems not too bad at all.

There was one other small change to the start-up message about the number
of CPUs available for threading.  Hopefully, the new message is clearer.