Eilmer3: House keeping for moving grid in Eilmer3
The current gas dynamic updating scheme with moving grid is predictor-corrector, I
was trying to speed up the turbulent simulation. If the cfl number of 1.0 is used,
this gives me the time step of 1.2e-10. However, a fixed time step of 5.0e-10
is also tried and the simulation remains stable. This commit is to comment out
cfl check after grid movement if fixed time step is configured.