Eilmer4: rebuild spatial derivative calculation via divergence theorem.
It now works on an arbitrary polygon in the xy plane, rather than assuming a
quadrilateral.

So far tested on the laminar flow along a flat plate.  Looks OK.
Still need to tidy up on redundant entries of corner points for cells in the
corners of each block but this doesn't affect function.
Also, still need to add options to select least-squares or divergence at
run time.