Eilmer3: added overshoot_factor to relax the extrema_clipping for Pierpaolo.
Hopefully, we should be able to reduce effective dissipation and so keep the
weak transverse waves in the expansion-tube simulations by turning off the
limiter but keeping the somewhat-relaxed extrema-clipping for numerical
stability.  Although it hasn't been tested much, the following settings appear
to work for the cone20 test.
gdata.apply_limiter_flag = 0
gdata.extrema_clipping_flag = 1
gdata.overshoot_factor = 1.5
A change to the solution can be seen, with slightly more extreme values.