Eilmer3: add new command-line option to report allowable timestep for all blocks.
-b or --report-dt-all-blocks

The intention is that the simulation can be run for short periods to gather
information on which block is dragging down the global time-step.
It should work in both shared-memory and MPI modes but the MPI mode may output
the blocks in an unordered fashion.  Hopefully, the interleaving of output
will not be any finer than a single line, otherwise the output will look really
garbled.  We will fix this if it becomes a problem.