Eilmer3: House keeping for the artificial dissipation term in Eilmer
It is recommended that only the fourth order artificial dissipation term is
actived during the simulation, this is to smooth the oscillation, while the
second order term is used in the vicinity of shock. For bearing simulation,
this can be turned off. The influence of the fourth order artificial dissipation
coefficient (0.01 and 0.02) has been done for foil bearings, the difference is
minor.