Eilmer3: Tidy up update of Menter omega correction made in Revision 2078
In Revision 2078, an update was made to implement the correct version of
Menter's omega correction on non-slip walls for the k-omega model. However,
the old version of Menter's correction was left commented out in the code.
In this commit, we remove all parts of the code that contains the old version
of Menter's omega correction.