lib/gas/kinetics: Implemented the handling of one-way reactions '=>' at the parser level.
Mengmeng Zhao (of USQ) had no success when using the documented feature of
the one-way reaction indicated by '=>' in a kinetics scheme. The bug was a rare
case of the documentation preceding the implementation. The one-way reaction
was not handled correctly by the parser; I had simply never had to handle that
case.

This fix will set up the reaction scheme such no backwards rates are included
when the one-way arrow is found in a chemical reaction.