lib/gas/kinetics: Bug fix in Reaction class so that one-way reactions are handled correctly.
A few days back, I corrected the chemistry pre-processor so that one-way reactions ('=>')
would be handled correctly. The pre-processor was doing its job correctly, but I hadn't
been careful that the internal implementation would also honour a one-way reaction.
This has now been fixed. I needed to make sure that backwards rate and equilibrium constant
pointers are all set to 0 in the case of a one-way reaction. This will force a zero
rate of production in the backwards direction.