Using the manufactured solutions from Roy's (2007) paper, the implementation of Wilcox's (2006) turbulence model in Eilmer3 has been successfully verified. After removing the bugs in source files, now we update the input files for turbulence model test.
1. Turbulence model can be tested uder different block arrangements, different grid
types and different flux calculation schemes.

2. For single-block arrangement, the observed order of accuracy should approach second order;
However, the observed order of spatial accuracy would reduce to first order for multi-block
arrangement, which is compatible with the results for viscous flows in Rowan's (2013) Eilmer paper.
The problem of dropping order of accuracy for viscous flows under multi-block arrangement is
expected to be solved in Eilmer4.