Eilmer3: MPI tasks to agree on stopping (or not) at the end of the main loop.
Each task is deciding it's own value of finished_time_stepping and, at the end
of the main iteration loop, all processors need to agree to proceed with another
iteration or to finish.

Pierre Mariotto was seeing some tasks stopping (on the max-wall-clock limit)
and others (presumably) not stopping and then waiting at the barrier at the top
of the next loop iteration.  His batch job would have a number of tasks write
the message that they were stopping and then the job would hang for an hour
or so, until the batch system killed it for exceeding the allowable time.
Hopefully, this revision will fix such bad program behaviour.