lib/cfpylib/gasdyn: change to normal_shock in gas_flow.py
This is a fairly large change to the normal_shock function in
gas_flow.py to make the function more robust. Most of the function
remains the same, but a maximum_iterations input has been added,
and an error will be thrown if the function does not converge
after this time. Some new stuff was added to the start to try
to better catch issues when the initial ideal gas guess gives
an unrealistic temperature, and the other function my_limiter
was reduced to a max change of 0.2 of the original value (from
0.5) to also make the normal_shock function more robust.

All of this was done under consultation with PJ and Rowan.

2 of the tests in test_gas_flow.py now fail, but will be fixed
in the morning.