lib/cfpylib/gasdyn: change to test_gas_flow.py to accomodate changes to gas_flow.py
after the changes committed yesterday to the normal_shock function in gas_flow.py
the pitot pressure in the test_expand_from_stagnation test was failing due to the delta
being too low. This has now been changed to be 1.0e-4*the pressure value

a similar problem was occuring in test_shock_given_V for p2 and this has also been changed
to be 1.0e-4* the test value.

This was all discussed with Peter earlier today as a reasonable solution