onedval: Added robustness to determining cell normal from Tecplot slice.
Onedval had only been used on slices that have normals that align
(mostly) with the x-axis. Based on this, I had a hack that accounted
for the fact that Tecplot will flip between an inward-facing and
an outward-facing normal in an indiscriminate manner when creating a
slice of cells.

This hack fell over for slices of more general orientation.
Luuk Meijboom identified this problem when applying onedval to
turbine passages. Our fix is to use the normal computed for the
very first cell in the slice and apply that everywhere. This
should be a very good approximation if Tecplot has indeed sliced
a true plane across the flow field.