app/pitot: a couple of very small pitot changes
found a print statement in pitot_condition_builder.py in the
normalised results printing function which was from testing and
didn't need to be there. has been removed.

dropped the minimum Vs1 range for the reflected shock tube conditions
in pitot so that it can be used for slower shock speeds. old minimum
was 1000 m/s which is a bit high