app/pitot: very small changes to pitot_output_utils and pitot_flow_functions
there was a bug in pitot_output_utils.py where if the stag_temp wasn't found,
it would try to print it anyway, causing the code to crash...

there was also a bug in pitot_flow_functions where if the two initial
Vs2 guesses in the acc tube failed, it would set them to guess_1 + 1000,
and guess_2 + 1000 and try again. This is fine, however guess_2 is already
guess_1 + 1000, so if guess_2 caused the failure, it will be the new
guess_2 and will fail again. Now the code sets them to guess_1 + 1500
and guess_2 + 1500 so this can't happen