app/pitot: made Pitot's eq mode work with CO2 based gases
Historically, I had used a pg-eq mode in the past, where the fill state
was set pg and I made the other states work in eq as they are
generally higher temperature.

The current solution sets the state 1 state using only the fill species (i.e.
say CO2 and N2) instead of all possible states, and it works much better.

This causes an issue with the calculation of Ht at the end of the pitot run
due to states 8 (an eq state) and state 1 (a pg state) not being
referenced to the same enthalpy reference temperature.

This required mainly changes to pitot_input_utils.py where the input states
are calculated using the function state_builder, and some small changes
to some of the flow functions in pitot_flow_functions.py

changes were also made to the function acceleration_tube_calculation
so that the composition of the accelerator gas is printed after running,
as this was seen to be useful as well...