app/pitot: found a bug with the csv output when a perfect gas driver is used
Turns out that I had made changes to the text file output to fix the issue,
but not the csv output. Rewrote the output for both the text and csv output
versions so that they are both a bit more better coded, and the csv one now
works for the perfect gas driver.