app/pitot: added ability set acc tube lower limit to Vs1
currently, the code would set the minimum Vs2 guess to a value slightly
above Vs1. This can be an issue for conditions which don't have a large
increase in shock speed from the shock to acceleration tubes. The lower
limit can be manually set lower, but it used to require a value,
which wouldn't work with situations like the condition builder, which
may do many different runs.  Iadded the ability make Vs2_lower = 'Vs1'
(a string) and then the code will pull out the Vs1 value from that
simulation for the lower limit. Also a small change to pitot_condition_builder.py
to make it work well with the condition builder.